% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{asc}
\alias{asc}
\alias{desc}
\title{Helpers to sort variables in ascending or descending order}
\usage{
asc(x)

desc(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A numeric vector that can be ordered in ascending or descending order. \cr
Useful in \code{dplyr::arrange()} or \code{f_arrange()}.
}
\description{
An alternative to \code{dplyr::desc()} which is much faster
for character vectors and factors.
}
