% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bind.R
\name{f_bind}
\alias{f_bind}
\alias{f_bind_rows}
\alias{f_bind_cols}
\title{Bind data frame rows and columns}
\usage{
f_bind_rows(...)

f_bind_cols(..., .repair_names = TRUE, .recycle = TRUE)
}
\arguments{
\item{...}{Data frames to bind.}

\item{.repair_names}{Should duplicate column names be made unique?
Default is \code{TRUE}.}

\item{.recycle}{Should inputs be recycled to a common row size?
Default is \code{TRUE}.}
}
\value{
\code{f_bind_rows()} performs a union of the data frames specified via \code{...} and
joins the rows of all the data frames, without removing duplicates.

\code{f_bind_cols()} joins the columns, creating unique column names if there are
any duplicates by default.
}
\description{
Faster bind rows and columns.
}
