% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_bind.R
\name{f_bind_rows}
\alias{f_bind_rows}
\alias{f_bind_cols}
\title{Bind data frame rows and columns}
\usage{
f_bind_rows(...)

f_bind_cols(..., .repair_names = TRUE, .sep = "...")
}
\arguments{
\item{...}{Data frames to bind.}

\item{.repair_names}{Should duplicate column names be made unique?
Default is \code{TRUE}.}

\item{.sep}{Separator to use for creating unique column names.}
}
\description{
Faster bind rows and columns.
}
