% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_expand.R
\name{f_expand}
\alias{f_expand}
\alias{f_complete}
\title{Fast versions of \code{tidyr::expand()} and \code{tidyr::complete()}.}
\usage{
f_expand(data, ..., sort = FALSE, .by = NULL)

f_complete(data, ..., sort = FALSE, .by = NULL, fill = NA)
}
\arguments{
\item{data}{A data frame}

\item{...}{Variables to expand}

\item{sort}{Logical. If \code{TRUE} expanded/completed variables are sorted.
The default is \code{FALSE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{fill}{A named list containing value-name pairs
to fill the named implicit missing values.}
}
\value{
A \code{data.frame} of expanded groups.
}
\description{
Fast versions of \code{tidyr::expand()} and \code{tidyr::complete()}.
}
