% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_filter.R
\name{f_filter}
\alias{f_filter}
\title{Alternative to \code{dplyr::filter()}}
\usage{
f_filter(data, ..., .by = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Expressions used to filter the data frame with.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}
}
\value{
A filtered data frame.
}
\description{
Alternative to \code{dplyr::filter()}
}
