% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ungroup.R
\name{f_ungroup}
\alias{f_ungroup}
\alias{group_ordered}
\title{Un-group \code{grouped_df}}
\usage{
f_ungroup(data)

group_ordered(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
An un-grouped data frame.
}
\description{
Un-group \code{grouped_df}
}
