% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastplyr-package.R
\docType{package}
\name{fastplyr-package}
\alias{fastplyr}
\alias{fastplyr-package}
\title{fastplyr: Fast Alternatives to 'tidyverse' Functions}
\description{
fastplyr is a tidy front-end using a faster and more efficient back-end based
on two packages, collapse and cheapr.

fastplyr includes dplyr and tidyr alternatives that behave like their
tidyverse equivalents but are more efficient.

Similar in spirit to the excellent tidytable package, fastplyr
also offers a tidy front-end that is fast and easy to use. Unlike tidytable,
fastplyr verbs are interchangeable with dplyr verbs.

You can learn more about the tidyverse, collapse and cheapr
using the links below.

\href{https://tidyverse.org/learn/}{tidyverse}

\href{https://fastverse.org/collapse/articles/collapse_intro.html}{collapse}

\href{https://github.com/NicChr/cheapr}{cheapr}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/fastplyr/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
\keyword{internal}
