% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastplyr_options.R
\name{fastplyr_options}
\alias{fastplyr_options}
\alias{fastplyr_enable_optimisations}
\alias{fastplyr_disable_optimisations}
\alias{fastplyr_enable_informative_msgs}
\alias{fastplyr_disable_informative_msgs}
\title{Setting global fastplyr options}
\usage{
fastplyr_enable_optimisations()

fastplyr_disable_optimisations()

fastplyr_enable_informative_msgs()

fastplyr_disable_informative_msgs()
}
\value{
Enables or disables fastplyr global options invisibly.
}
\description{
Helper functions to allow users to:
\itemize{
\item Enable or disable optimisations for common functions package-wide
\item Enable or disable informative messages
}
}
\seealso{
\link{get_group_unaware_fns}
}
