% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_unaware.R
\name{get_group_unaware_fns}
\alias{get_group_unaware_fns}
\title{Get list of current group-unaware functions}
\usage{
get_group_unaware_fns()
}
\value{
A named list of functions marked as group-unaware in fastplyr.
}
\description{
Get list of current group-unaware functions
}
\examples{
library(fastplyr)

fns <- get_group_unaware_fns()

names(fns)
fns$round

}
