% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_order_default.R
\name{group_by_order_default}
\alias{group_by_order_default}
\title{Default value for ordering of groups}
\usage{
group_by_order_default(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
A default value, \code{TRUE} or \code{FALSE} that controls which algorithm to use
for calculating groups. See \link{f_group_by} for more details.
}
