% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_group_data.R
\name{group_data}
\alias{group_data}
\alias{f_group_data}
\alias{f_group_keys}
\alias{f_group_rows}
\alias{f_group_indices}
\alias{f_group_vars}
\alias{f_group_size}
\alias{f_n_groups}
\title{Fast group metadata}
\usage{
f_group_data(x)

f_group_keys(x)

f_group_rows(x)

f_group_indices(x)

f_group_vars(x)

f_group_size(x)

f_n_groups(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{grouped_df}.}
}
\value{
Requested group metadata.
}
\description{
Fast group metadata
}
