% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tidy.R
\name{list_tidy}
\alias{list_tidy}
\title{Alternative to \code{rlang::list2}}
\usage{
list_tidy(..., .keep_null = TRUE, .named = FALSE)
}
\arguments{
\item{...}{Dynamic name-value pairs.}

\item{.keep_null}{\verb{[logical(1)]} - Should \code{NULL} elements be kept?
Default is \code{TRUE}.}

\item{.named}{\verb{[logical(1)]} - Should all list elements be named?
Default is \code{FALSE}.}
}
\description{
Evaluates arguments dynamically like \code{rlang::list2} but objects
created in \code{list_tidy} have precedence over environment objects.
}
