% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisations.R
\name{optimisations}
\alias{optimisations}
\alias{fastplyr_enable_optimisations}
\alias{fastplyr_disable_optimisations}
\title{Enable and disable fastplyr optimisations}
\usage{
fastplyr_enable_optimisations()

fastplyr_disable_optimisations()
}
\value{
Enables or disables optimisations invisibly.
}
\description{
Enable or disable optimisations for common functions package-wide.
}
\seealso{
\link{get_group_unaware_fns}
}
