% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_rows_if_any_na}
\alias{remove_rows_if_any_na}
\alias{remove_rows_if_all_na}
\title{Fast remove rows with \code{NA} values}
\usage{
remove_rows_if_any_na(.data, ..., .cols = NULL)

remove_rows_if_all_na(.data, ..., .cols = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{...}{Cols to fill \code{NA} values specified through \code{tidyselect} notation.
If left empty all cols are used by default.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}
}
\value{
A data frame with removed rows containing either any or all \code{NA} values.
}
\description{
Fast remove rows with \code{NA} values
}
