% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastplyr.R
\docType{package}
\name{fastplyr-package}
\alias{fastplyr}
\alias{fastplyr-package}
\title{fastplyr: Fast Alternatives to 'dplyr' Functions}
\description{
Alternatives that behave like their tidyverse equivalents but are more efficient.

You can learn more about the tidyverse and collapse using the links below

\href{https://www.tidyverse.org/learn/}{tidyverse}

\href{https://sebkrantz.github.io/collapse/articles/collapse_intro.html}{collapse}
}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
