% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{across}
\alias{pick}
\alias{n}
\alias{all_of}
\alias{any_of}
\alias{where}
\alias{contains}
\alias{starts_with}
\alias{ends_with}
\alias{everything}
\alias{num_range}
\alias{last_col}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{across}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{pick}}}

  \item{tidyselect}{\code{\link[tidyselect]{all_of}}, \code{\link[tidyselect:all_of]{any_of}}, \code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{where}}}
}}

